requirements.txt

Flask==3.0.3
python-dotenv==1.0.1
ollama==0.2.1
gTTS==2.5.1